--- apps/test-app/src/frontend/UrlParams.tsx
@@ -1,28 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-import * as React from "react";
-import { PreviewFeatures } from "@itwin/appui-react";
-
-export function usePreviewFeatureURLParams() {
-  return React.useMemo(() => {
-    const reparentPopoutWidgets = getUrlParam("reparentPopoutWidgets");
-    const previewFeatures: PreviewFeatures = {};
-    if (reparentPopoutWidgets) {
-      previewFeatures.reparentPopoutWidgets = Array.isArray(
-        reparentPopoutWidgets
-      )
-        ? reparentPopoutWidgets
-        : reparentPopoutWidgets !== "0";
-    }
-    return previewFeatures;
-  }, []);
-}
-
-export function getUrlParam(name: string) {
-  const url = new URL(window.location.href);
-  const params = new URLSearchParams(url.search);
-  const param = params.get(name);
-  return param === null ? undefined : param;
-}
